devel/llvm-devel: New snapshot
Disable FLANG on arm6/arm7.  It's not 32-bit clean, is unlikely to
ever be so, and fails to build on many people's systems due to running
out of memory.  This matches other llvm-* ports.

Replace a patch reverting a commit that is broken with our libc++ with a
patch with a workaround that is likely to land upstream.